requirements.txt
streamlit
groq
pymupdf
python-docx
python-dotenv
from-root
mangum
-e .
